Cleans up tests.
